third_party/WebKit/LayoutTests/pointer-lock/mouse-event-api.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
    var mouseEvent = document.createEvent("MouseEvent");
    shouldBeDefined("mouseEvent.movementX");
    shouldBeDefined("mouseEvent.movementY");
</script>
</body>
</html>
